sort tests into suites
